Remove some older functions, add replace for array
This commit removed the older functions "cjson_push_child" and
"cjson_push_item". You should instead use "cjson_append" and
"cjson_insert".

There have also been some more additions of documentation
in the cjson header file.

There is also a function for replacing an element in an array,
similar to the replace function of an object.